feat(forge): raise sequence-array batch limit and rewrite the well API docs
- Raise the Molecular Sequence Rewrite Array and Assembler Matrix Sequence
  Rewrite Core logical batch limit from Integer.MAX_VALUE to Long.MAX_VALUE;
  multiplication and buffer overflow are rejected without consuming inputs.
- Add the controller's pattern-only item interface for ME storage buses and
  item automation, sync pattern edits per crystal, and keep explicit pattern
  ownership on recovery and dismantling.
- Rewrite docs/matter-research-api.md for the 1.20.1 Forge port in English,
  add a cross-linked Chinese version (docs/matter-research-api.zh-CN.md) that
  documents the recipes/ directory, forge:conditions, ForgeRecipeCodecs and
  the recipe-object research API, and update the API index, both READMEs,
  the batch-provider cross-link and the in-game guide.
- Refresh CHANGELOG, 2.0.3 release notes and the CurseForge description with
  the new validation results.

+++ src/main/java/com/atir/molecularmanipulator/blockentity/AssemblerMatrixMolecularCoreBlockEntity.java
@@ -33,7 +33,7 @@ import net.minecraft.world.level.block.state.BlockState;
 import java.util.List;
 
 public final class AssemblerMatrixMolecularCoreBlockEntity extends TileAssemblerMatrixFunction implements IGridTickable {
-    public static final long VIRTUAL_PARALLEL_LIMIT = Integer.MAX_VALUE;
+    public static final long VIRTUAL_PARALLEL_LIMIT = Long.MAX_VALUE;
     private static final int MAX_BUFFERED_TYPES = 256;
     private static final String OUTPUT_BUFFER_TAG = "output_buffer";
     private static final String OUTPUT_READY_TICK_TAG = "output_ready_tick";
